+++ NEWS.md
@@ -1,3 +1,11 @@
+pmsampsize v1.0.2 (Release date: 2019-11-15)
+=============================================
+
+* update forces the time-unit to be the mean follow-up time, as described in forthcoming BMJ paper which extends the original Statistics in Medicine paper (see reference in help file)
+
+* updated in calculation of max R-sq for survival option, to prevent manipulation based on time units used 
+
+
 pmsampsize v1.0.1 (Release date: 2019-08-02)
 =============================================